// src/types/LeaseDocExtractedTerm.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type LeaseDocExtractedTerm = {
    leaseDocExtractedTermID: string;
    leaseDocumentID: string;
    extractedStartDate: string;
    extractedEndDate: string;
    extractedRentAmount: number;
    extractedDepositAmount: number;
    extractedRenewalDate: string;
    extractionConfidence: number;
    extractionStatus: string;
    rawExtractionJson: string;
    extractedDate: string;
    reviewedAt: string;
    reviewedBy: string;
};